libroach: move utility methods to dbhelper.cc
This is a) an attempt to down-size `db.cc` at least a little bit (though
that file remains at thousands of lines) and b) was useful for moving
`mvccScanner` out of `db.cc`.

Release note: None
