+++ c-deps/libroach/db.h
@@ -13,6 +13,7 @@
 // permissions and limitations under the License.
 
 #include <libroach.h>
+#include <memory>
 #include <rocksdb/comparator.h>
 #include <rocksdb/iterator.h>
 #include <rocksdb/status.h>
@@ -54,3 +55,14 @@ const ::rocksdb::Comparator* CockroachComparator();
 // Stats are only computed for keys between the given range.
 MVCCStatsResult MVCCComputeStatsInternal(::rocksdb::Iterator* const iter_rep, DBKey start,
                                          DBKey end, int64_t now_nanos);
+
+struct DBIterator {
+  std::unique_ptr<rocksdb::Iterator> rep;
+  std::unique_ptr<rocksdb::WriteBatch> kvs;
+  std::unique_ptr<rocksdb::WriteBatch> intents;
+};
+
+std::string ToString(DBSlice s);
+std::string ToString(DBString s);
+rocksdb::Slice ToSlice(DBSlice s);
+rocksdb::Slice ToSlice(DBString s);
